pyproject.toml
[project]
name = "ghlens"
version = "0.1.0"
description = "Fetch PR metadata and comments from the GitHub GraphQL API"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "rich>=13.7",
]

[project.scripts]
ghlens = "ghlens.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ghlens"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
    "respx>=0.21",
]
